+++ bisheng.config.js
@@ -21,6 +21,31 @@ const path = require('path');
 //       amd: 'moment',
 //     };
 //   }
+const configList = [
+    {
+    loader: 'babel-loader',
+    options: {
+        // 不采用.babelrc的配置
+        "babelrc": false,
+        "presets": [
+            ["react"],
+            ["es2015", { "modules": false }]
+        ],
+        "plugins": [
+            "syntax-dynamic-import",
+            "transform-object-rest-spread",
+            "transform-class-properties"
+        ]
+    }
+},
+{
+    test: /\.(eot|woff|svg|ttf|woff2|gif|appcache|webp)(\?|$)/,
+    loader: [
+        "file-loader?name=[name].[ext]",
+        "url-loader?limit=100000"
+    ]
+}
+]
 module.exports = {
     target: 'node',
     source: {
@@ -52,13 +77,7 @@ module.exports = {
     webpackConfig: function (config) {
         // ignoreMomentLocale(config);
         // externalMoment(config);
-        config.module.rules.push({
-            test: /\.(eot|woff|svg|ttf|woff2|gif|appcache|webp)(\?|$)/,
-            loader: [
-                "file-loader?name=[name].[ext]",
-                "url-loader?limit=100000"
-            ]
-        });
+        config.module.rules.concat(configList);
         config.optimization = {
             splitChunks: {
                 chunks: "all",
