--- components/calendar/calend.tsx
@@ -1,8 +1,0 @@
-import { Dayjs } from 'dayjs';
-import dayjsGenerateConfig from 'rc-picker/es/generate/dayjs';
-import generateCalendar from 'antd/es/calendar/generateCalendar';
-import 'antd/es/calendar/style';
-
-const Calendar = generateCalendar<Dayjs>(dayjsGenerateConfig);
-
-export default Calendar;
