+++ components/timepicker/tmpicker.js
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import DatePicker from '../datePicker/dtpicker';
+
+const TimePicker = React.forwardRef((props, ref) => {
+  return <DatePicker {...props} picker="time" mode={undefined} ref={ref} />;
+});
+
+TimePicker.displayName = 'TimePicker';
+
+export default TimePicker;
