--- tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-      "outDir": "dist", //重定向输出目录。
-      "target": "es5",
-      "lib": ["esnext", "dom"],
-      "sourceMap": true,
-      "allowJs": true,
-      "removeComments": false,
-      "baseUrl": ".",
-      "jsx": "react",
-      "importHelpers": true,
-      "skipLibCheck": true,
-      "allowSyntheticDefaultImports": true,
-      "esModuleInterop": true,
-      "forceConsistentCasingInFileNames": true,
-      "experimentalDecorators": true,
-      "downlevelIteration": true,
-      "paths": {
-        "types/*": ["./types/*"]
-      }
-    },
-    "include": ["components/**/*.tsx", "node_modules/dt-common/**/*"]
-  }
-
